Refactor project to allow proper builds & add self-contained flag.
