+++ Backend/Backend.csproj
@@ -6,7 +6,6 @@
     </PropertyGroup>
 
     <ItemGroup>
-      <PackageReference Include="BenchmarkDotNet" Version="0.13.1" />
       <PackageReference Include="BetterConsoleTables" Version="2.0.4-rc1" />
     </ItemGroup>
