+++ StockNemo.sln
@@ -6,6 +6,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Terminal", "Terminal\Termin
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Test", "Test\Test.csproj", "{2DB6E690-8290-4FFC-9A6E-F4E7F9630B21}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Benchmark", "Benchmark\Benchmark.csproj", "{22D02EA8-275D-46DB-A5E5-235AD8A55602}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -24,5 +26,9 @@ Global
 		{2DB6E690-8290-4FFC-9A6E-F4E7F9630B21}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{2DB6E690-8290-4FFC-9A6E-F4E7F9630B21}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{2DB6E690-8290-4FFC-9A6E-F4E7F9630B21}.Release|Any CPU.Build.0 = Release|Any CPU
+		{22D02EA8-275D-46DB-A5E5-235AD8A55602}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{22D02EA8-275D-46DB-A5E5-235AD8A55602}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{22D02EA8-275D-46DB-A5E5-235AD8A55602}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{22D02EA8-275D-46DB-A5E5-235AD8A55602}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
