+++ Terminal/Terminal.csproj
@@ -17,8 +17,4 @@
       <ProjectReference Include="..\Backend\Backend.csproj" />
     </ItemGroup>
 
-    <ItemGroup>
-      <PackageReference Include="BenchmarkDotNet" Version="0.13.1" />
-    </ItemGroup>
-
 </Project>
